+++ Assets/InputSystem/InputSettings.asset
@@ -10,7 +10,7 @@ MonoBehaviour:
   m_Enabled: 1
   m_EditorHideFlags: 0
   m_Script: {fileID: 11500000, guid: c46f07b5ed07e4e92aa78254188d3d10, type: 3}
-  m_Name: StarterAssets.inputsettings
+  m_Name: InputSettings
   m_EditorClassIdentifier: 
   m_SupportedDevices: []
   m_UpdateMode: 1
